database config updated

+++ app/routes/employee.py
@@ -13,7 +13,7 @@ router = APIRouter(
 
 # list = []
 
-
+# create employee
 @router.post("/", status_code=status.HTTP_201_CREATED)
 async def create_employee(employee: schemas.EmployeeCreate, db: Session = Depends(get_db)):
     new_employee = models.Employee(**employee.model_dump())
@@ -30,12 +30,16 @@ async def create_employee(employee: schemas.EmployeeCreate, db: Session = Depend
 
     return new_employee
 
+# get all employees
+
 
 @router.get("/")
 def get_employees(db: Session = Depends(get_db)):
     employees = db.query(models.Employee).all()
 
     return employees
+
+# get employee by id
 
 
 @router.get("/{id}", response_model=schemas.EmployeeOut)
@@ -52,20 +56,26 @@ def get_employee(id: int, db: Session = Depends(get_db)):
     return employee
 
 
+# delete employee by id
 @router.delete("/{id}")
 def delete_employee(id: int, db: Session = Depends(get_db)):
     employee = db.query(models.Employee).filter(
         models.Employee.id == id).first()
+
     employee_name = db.query(models.Employee.name).first()
+
+    if not employee:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"{employee_name} does not exist")
     db.delete(employee)
     db.commit()
 
     return {"message": f"Employee {employee_name} has been deleted"}
 
 
-#update employee by id
+# update employee by id
 @router.put("/{id}")
-def update_employee(id: int, update_data: Dict[str, Any], db: Session = Depends(get_db)):
+async def update_employee(id: int, update_data: Dict[str, Any], db: Session = Depends(get_db)):
     existing_employee = db.query(models.Employee).filter(
         models.Employee.id == id).first()
     if not existing_employee:
@@ -78,8 +88,3 @@ def update_employee(id: int, update_data: Dict[str, Any], db: Session = Depends(
     db.commit()
 
     return {"message": "Employee has been updated"}
-
-# @router.put("/{id}")
-# def update_employee(id: int, employee: schemas.EmployeeUpdate, db: Session = Depends(get_db)):
-#     employee = db.query(models.Employee).filter(
-#         models.Employee.id == id).first()
